Update Snake and TwoEleven

+++ Apps/TwoEleven/main/Source/TwoElevenLogic.c
@@ -126,7 +126,7 @@ bool game_over(uint16_t matrix_size, const uint16_t **matrix) {
 /**
  * @brief Get the current score
  */
-uint16_t twoeleven_get_score(lv_obj_t * obj)
+uint32_t twoeleven_get_score(lv_obj_t * obj)
 {
     const twoeleven_t * game_2048 = (const twoeleven_t *)lv_obj_get_user_data(obj);
     if (!game_2048) return 0;
